% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{fuse.cluster}
\alias{fuse.cluster}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
fuse.cluster(K0, K1, sort = TRUE)
}
\arguments{
\item{K0}{Integer or numeric matrix with unmethylated counts}

\item{K1}{Integer or numeric matrix with methylated counts}

\item{sort}{Default is true, sorts the tree in optimal order}
}
\value{
Clustering tree as a matrix
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
\examples{
# Example: Clustering generated data
K0 <- matrix(sample(c(1:200), 125, replace = TRUE), ncol = 5)
K1 <- matrix(sample(c(1:200), 125, replace = TRUE), ncol = 5)
tree <- fuse.cluster(K0, K1)
tree

}
