% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{fuse.sort.tree}
\alias{fuse.sort.tree}
\title{Sort Clustering Tree in Optimal Order}
\usage{
fuse.sort.tree(tree)
}
\arguments{
\item{tree}{Unsorted clustering tree}
}
\value{
Sorted tree
}
\description{
Sorts the clustering tree produced by \code{\link{fuse.cluster}}.
}
\examples{
# Example: Sorting unsorted clustering tree
K0 <- matrix(sample(c(1:200), 80, replace = TRUE), ncol = 5)
K1 <- matrix(sample(c(1:200), 80, replace = TRUE), ncol = 5)
tree <- fuse.cluster(K0, K1, sort = FALSE)
tree.sorted <- fuse.sort.tree(tree)
tree

}
