% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{number.of.clusters}
\alias{number.of.clusters}
\title{Find Optimal Number of Clusters}
\usage{
number.of.clusters(tree, n, method = c("BIC", "AIC"))
}
\arguments{
\item{tree}{A sorted clustering tree (matrix or data.frame).}

\item{n}{Number of samples in the original data.}

\item{method}{Information criterion method. One of \code{"BIC"} or \code{"AIC"}.}
}
\value{
An integer representing the optimal number of clusters.
}
\description{
Determines the optimal number of clusters to cut a hierarchical clustering tree,
based on the selected information criterion (e.g., BIC or AIC).
}
\examples{
# Example: Determine number of clusters in dummy data set
K0 <- matrix(sample(1:200, 80, replace = TRUE), ncol = 5)
K1 <- matrix(sample(1:200, 80, replace = TRUE), ncol = 5)
tree <- fuse.cluster(K0, K1, sort = TRUE)
k <- number.of.clusters(tree, ncol(K0), 'BIC')
k

}
