% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.fuse_summary}
\alias{plot.fuse_summary}
\title{Plot method for FUSE segmentation results}
\usage{
\method{plot}{fuse_summary}(x, ..., segments_to_plot = 1:50)
}
\arguments{
\item{x}{A fuse_summary object}

\item{...}{Additional arguments}

\item{segments_to_plot}{Integer vector of segment indices}
}
\description{
Plot method for FUSE segmentation results
}
