% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{fuse.cluster}
\alias{fuse.cluster}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
fuse.cluster(K0, K1)
}
\arguments{
\item{K0}{Integer or numeric matrix with unmethylated counts}

\item{K1}{Integer or numeric matrix with methylated counts}
}
\value{
Clustering tree as a matrix, with the same structure as hclust, including the columns
\describe{
\item{m1}{ID of first data point. <0 is original point, >0 is row of previous merge.}
\item{m2}{ID of second data point. <0 is original point, >0 is row of previous merge.}
\item{logl_tot}{Change in total log-likelihood of merge.}
\item{logl_merge}{Total cost of points in this merge.}
\item{genomic_dist}{Penalty for genomic distance between points.}
}
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
\examples{
# Example: Clustering generated data
set.seed(1234)
K0 <- matrix(
  rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
K1 <- matrix(
  rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)
tree

}
