% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-cluster.R, R/methods-bioconductor.R
\name{fuse.cluster}
\alias{fuse.cluster}
\alias{fuse.cluster.default}
\alias{fuse.cluster.BSseq}
\alias{fuse.cluster.methrix}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
fuse.cluster(x, ...)

\method{fuse.cluster}{default}(x, K1, chr = NULL, pos = NULL, ...)

\method{fuse.cluster}{BSseq}(x, ...)

\method{fuse.cluster}{methrix}(x, ...)
}
\arguments{
\item{x}{Input object. One of:
\describe{
\item{matrix}{Unmethylated count matrix (K0).}
\item{BSseq}{A \code{BSseq} object.}
\item{methrix}{A \code{methrix} object.}
}}

\item{...}{Additional arguments if K0 is a matrix.}

\item{K1}{Methylated count matrix (if x is matrix).}

\item{chr}{Chromosome labels (if x is matrix).}

\item{pos}{Genomic positions (if x is matrix.}
}
\value{
A clustering tree of class \code{hclust}.
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
\examples{
# Example: Clustering generated data
set.seed(1234)
K0 <- matrix(
  rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
K1 <- matrix(
  rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)
tree

}
