% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse.cluster}
\alias{fuse.cluster}
\title{Perform Hierarchical Clustering on Methylation Data}
\usage{
fuse.cluster(K0, K1, sort = TRUE)
}
\arguments{
\item{K0}{Integer matrix with unmethylated counts}

\item{K1}{Integer matrix with methylated counts}

\item{sort}{Default is true, sorts the tree in optimal order}
}
\value{
Clustering tree as a matrix
}
\description{
Produces a hierarchical clustering tree based on the input matrices of counts.
}
