% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-cut.R
\name{fuse.cut.tree}
\alias{fuse.cut.tree}
\title{Cut Hierarchical Clustering Tree into Clusters}
\usage{
fuse.cut.tree(tree, k)
}
\arguments{
\item{tree}{Clustering tree of class \code{hclust}.}

\item{k}{Number of clusters}
}
\value{
A vector indicating which cluster each element in the original data frame belonged to
}
\description{
Divides the clustering tree into a specified number of clusters.
}
\examples{
# Example: Cutting small tree in 2 segments
set.seed(1234)
K0 <- matrix(
 rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
 nrow = 100, byrow = TRUE
)
K1 <- matrix(
 rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
 nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)

segments <- fuse.cut.tree(tree, 2)
segments

}
