% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse.sort.tree}
\alias{fuse.sort.tree}
\title{Sort Clustering Tree in Optimal Order}
\usage{
fuse.sort.tree(tree)
}
\arguments{
\item{tree}{Unsorted clustering tree}
}
\value{
Sorted tree
}
\description{
Sorts the clustering tree produced by \code{\link{fuse.cluster}}.
}
