% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse-cut.R
\name{number.of.clusters}
\alias{number.of.clusters}
\title{Find Optimal Number of Clusters}
\usage{
number.of.clusters(tree, n, method = c("BIC", "AIC"))
}
\arguments{
\item{tree}{Clustering tree of class \code{hclust}.}

\item{n}{Number of samples in the original data.}

\item{method}{Information criterion method. One of \code{"BIC"} or \code{"AIC"}.}
}
\value{
An integer representing the optimal number of clusters.
}
\description{
Determines the optimal number of clusters to cut a hierarchical clustering tree,
based on the selected information criterion (e.g., BIC or AIC).
}
\examples{
# Example: Determine number of clusters in dummy data set
set.seed(1234)
K0 <- matrix(
  rep(c(sample(0:20, 200, replace = TRUE), sample(20:40, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
K1 <- matrix(
  rep(c(sample(20:40, 200, replace = TRUE), sample(0:20, 200, replace = TRUE)), 2),
  nrow = 100, byrow = TRUE
)
tree <- fuse.cluster(K0, K1)
k <- number.of.clusters(tree, ncol(K0), 'BIC')
k

}
